+++ component/loki/source/file/file_test.go
@@ -76,6 +76,63 @@ func Test(t *testing.T) {
 	}
 }
 
+func TestBackoff(t *testing.T) {
+	defer goleak.VerifyNone(t, goleak.IgnoreTopFunction("go.opencensus.io/stats/view.(*worker).start"))
+	ctx, cancel := context.WithCancel(componenttest.TestContext(t))
+
+	// Create file to log to.
+	f, err := os.CreateTemp(t.TempDir(), "example")
+	require.NoError(t, err)
+	defer f.Close()
+
+	ctrl, err := componenttest.NewControllerFromID(util.TestLogger(t), "loki.source.file")
+	require.NoError(t, err)
+
+	ch1 := loki.NewLogsReceiver()
+
+	args := Arguments{
+		Targets: []discovery.Target{{
+			"__path__": f.Name(),
+			"foo":      "bar",
+		}},
+		ForwardTo: []loki.LogsReceiver{ch1},
+		Backoff: Backoff{
+			MinBackoff: time.Millisecond * 500,
+			MaxBackoff: time.Millisecond * 500,
+		},
+	}
+
+	go func() {
+		err := ctrl.Run(ctx, args)
+		require.NoError(t, err)
+	}()
+
+	ctrl.WaitRunning(time.Minute)
+
+	timeBeforeWriting := time.Now()
+
+	// Sleep for 600ms to miss the first poll, the next poll should be 2*MinBackoff later.
+	time.Sleep(time.Millisecond * 600)
+
+	_, err = f.Write([]byte("writing some text\n"))
+	require.NoError(t, err)
+
+	select {
+	case logEntry := <-ch1.Chan():
+		require.Greater(t, time.Now().Sub(timeBeforeWriting), 1*time.Second)
+		require.WithinDuration(t, time.Now(), timeBeforeWriting, 2*time.Second)
+		require.Equal(t, "writing some text", logEntry.Line)
+	case <-time.After(5 * time.Second):
+		require.FailNow(t, "failed waiting for log line")
+	}
+
+	// Shut down the component.
+	cancel()
+
+	// Wait to make sure that all go routines stopped.
+	time.Sleep(time.Millisecond * 100)
+}
+
 // Test that updating the component does not leak goroutines.
 func TestUpdate_NoLeak(t *testing.T) {
 	defer goleak.VerifyNone(t, goleak.IgnoreTopFunction("go.opencensus.io/stats/view.(*worker).start"))